v1.1.0
Added core file extender since the needed event while adding the product does only get triggered if the product does have customfields.

Now we create an own event with the modification of the cart.php Virtuemart file.
The modification will be automatically re-added if Virtuemart gets updated and can be readded with the plugin parameters.

Additionally the plugin get's automatically enabled after the installation and the installation routine has checks for the needed PHP and Joomla Versions.